requirements.txt
cloudinary==1.36.0
python-dotenv==1.0.0
supabase==2.8.1 